#369: Do not gzip video files when deploying frontend files into s3

+++ src/lib/Property/Frontend.js
@@ -263,6 +263,16 @@ export class Frontend {
       throw new FailedUploadingFileToS3Exception('*', bucketName, syncResultHtml.error);
     }
 
+    console.debug(`Syncing ${this.path} with ${bucketName} (Video only, TTL=86400)`);
+
+    let syncResultVideos = this
+      ._getSyncCommandVideoFiles(credentialsFile, bucketName, bucketRegion)
+      .runSync();
+
+    if (syncResultVideos.failed) {
+      throw new FailedUploadingFileToS3Exception('*', bucketName, syncResultVideos.error);
+    }
+
     FileSystem.unlinkSync(credentialsFile);
 
     // @todo: improve this by using directory upload
@@ -291,7 +301,6 @@ export class Frontend {
     return syncStack.join();
   }
 
-
   /**
    * @param {String} credentialsFile
    * @param {String} bucketName
@@ -300,6 +309,12 @@ export class Frontend {
    * @private
    */
   _getSyncCommandNoHtml(credentialsFile, bucketName, bucketRegion) {
+    let excludesStr = '--exclude="*.html" '; // exclude *.html by default
+
+    Frontend._videoAssetsExtensions.forEach(extension => {
+      excludesStr += `--exclude="${extension}" `;
+    });
+
     return new Exec(
       `export AWS_CONFIG_FILE=${credentialsFile};`,
       'aws s3 sync',
@@ -308,7 +323,8 @@ export class Frontend {
       '--storage-class=REDUCED_REDUNDANCY',
       Frontend._contentEncodingExecOption,
       '--cache-control="max-age=86400"',
-      '--exclude="*.html"',
+      '--include="*"',
+      excludesStr,
       `'${this.path}'`,
       `'s3://${bucketName}'`
     );
@@ -335,6 +351,41 @@ export class Frontend {
       `'${this.path}'`,
       `'s3://${bucketName}'`
     );
+  }
+
+  /**
+   * @param {String} credentialsFile
+   * @param {String} bucketName
+   * @param {String} bucketRegion
+   * @returns {Exec}
+   * @private
+   */
+  _getSyncCommandVideoFiles(credentialsFile, bucketName, bucketRegion) {
+    let includesStr = '';
+    Frontend._videoAssetsExtensions.forEach(extension => {
+      includesStr += `--include="${extension}" `;
+    });
+
+    return new Exec(
+      `export AWS_CONFIG_FILE=${credentialsFile};`,
+      'aws s3 sync',
+      '--profile=deep',
+      `--region=${bucketRegion}`,
+      '--storage-class=REDUCED_REDUNDANCY',
+      '--cache-control="max-age=86400"',
+      '--exclude="*"',
+      includesStr,
+      `'${this.path}'`,
+      `'s3://${bucketName}'`
+    );
+  }
+
+  /**
+   * @returns {String[]}
+   * @private
+   */
+  static get _videoAssetsExtensions() {
+    return ['*.avi', '*.fvl', '*.mp4', '*.wmv', '*.mov'];
   }
 
   /**
